utitlityFunction
